+++ test/tmp-path-test.js
@@ -1,0 +1,3 @@
+const path = require('path');
+console.log(JSON.stringify(path.join('C:\\project', '.workspace-bridge')));
+console.log(JSON.stringify(path.join('c:/project', '.workspace-bridge')));